+++ just_audio/darwin/just_audio/Sources/just_audio/BetterEventChannel.m
@@ -32,7 +32,11 @@
 
 - (void)dispose {
     if (_eventSink) {
-        _eventSink(FlutterEndOfEventStream);
+        @try {
+            _eventSink(FlutterEndOfEventStream);
+        } @catch (NSException *exception) {
+            NSLog(@"Exception while ending event stream: %@", exception.reason);
+        }
     }
     [_eventChannel setStreamHandler:nil];
 }