+++ just_audio/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 0.11.0
 
+## 0.10.4
+
+* Fix bug on simultaneous loads.
+* Fix dispose crash on iOS.
+* Fix initial seek values on subsequent loads (@Abestanis).
+
 ## 0.10.3
 
 * Fix pending timers bug in unit tests.